commented out code that is moved to ansible or not used
